requirements.txt
Scrapy
arrow
google-cloud-pubsub==0.39.1
ipdb
beautifulsoup4
html5lib
